Renamed api/handlers to api/channels
